requirements.txt
accelerate==0.32.0
aiofiles==23.2.1
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
audioread==3.0.1
auditok==0.1.5
Babel==2.15.0
beautifulsoup4==4.12.3
bibtexparser==2.0.0b7
blinker==1.8.2
cachetools==5.3.3
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clldutils==3.22.2
colorama==0.4.6
colorlog==6.8.2
comtypes==1.4.5
contourpy==1.2.1
cssselect2==0.7.0
csvw==3.3.0
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.8
dlinfo==1.2.1
dnspython==2.6.1
docopt==0.6.2
einops==0.8.0
einops-exts==0.0.4
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
faust-cchardet==2.1.19
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.37.2
gradio_client==1.0.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_resources==6.4.0
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
language-tags==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
monotonic_align @ git+https://github.com/resemble-ai/monotonic_align.git@78b985be210a03d08bc3acc01c4df0442105366f
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
munch==4.0.0
networkx==3.3
nltk==3.8.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
orjson==3.10.6
packaging==24.1
pandas==2.2.2
phonemizer==3.2.1
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
plum-dispatch==1.7.4
pooch==1.8.2
proglog==0.1.10
protobuf==5.27.2
psutil==6.0.0
pyarrow==16.1.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pylatexenc==2.10
pyparsing==3.1.2
pysrt==1.1.2
pysubs2==1.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
pyvips==2.2.3
PyYAML==6.0.1
ray==2.32.0
rdflib==7.0.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.1
ruff==0.5.0
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.5.1
scipy==1.14.0
segments==2.2.1
semantic-version==2.10.0
sentencepiece==0.1.99
setuptools
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
sox==1.5.0
soxr==0.3.7
srt==3.5.3
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision
tornado==6.4.1
tqdm==4.66.4
transformers==4.42.3
triton==2.3.1
txtsplit==1.0.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
unotools==0.3.3
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.3
watchdog==4.0.1
watchfiles==0.22.0
webencodings==0.5.1
webrtcvad==2.0.10
websockets==11.0.3
XlsxWriter==3.2.0
